feat(#103 toast): asset에 토스트 추가

+++ apps/v1/src/admin/pages/Asset/Management.tsx
@@ -1,43 +1,80 @@
-import { Button, DataTable, Skeleton } from '@plug/ui';
-import { columns } from './constants/assetColumns';
-import { AssetRegistModal } from '../../components/modals/AssetRegist'; 
-import { useModal } from '../../components/hook/useModal';
-import { useAssetsSWR, deleteAsset } from '@plug/common-services';
-import { useAsset } from './utils/useAsset';
-import { StateInfoWrapper } from "@plug/v1/admin/components/boundary/StateInfoWrapper";
-import { useState } from 'react';
-import { Asset } from './types/asset.types';
+import {Button, DataTable, Skeleton} from '@plug/ui';
+import {columns} from './constants/assetColumns';
+import {AssetRegistModal} from '../../components/modals/AssetRegist';
+import {useModal} from '../../components/hook/useModal';
+import {useAssetsSWR, deleteAsset} from '@plug/common-services';
+import {useAsset} from './utils/useAsset';
+import {StateInfoWrapper} from "@plug/v1/admin/components/boundary/StateInfoWrapper";
+import {useState} from 'react';
+import {Asset} from './types/asset.types';
+import {useToastStore} from "@plug/v1/admin/components/hook/useToastStore";
 
 export default function AssetPage() {
-    const { isOpen, mode, openModal, closeModal } = useModal();
-    const { data, error, isLoading, mutate } = useAssetsSWR();
-    const [ selectedAssets, setSelectedAssets ] = useState<Set<Asset>>(new Set());
-    const [ selectedAssetId, setSelectedAssetId ] = useState<number>();
-    
+    const {isOpen, mode, openModal, closeModal} = useModal();
+    const {data, error, isLoading, mutate} = useAssetsSWR();
+
+    const [selectedAssets, setSelectedAssets] = useState<Set<Asset>>(new Set());
+    const [selectedAssetId, setSelectedAssetId] = useState<number>();
+
+    const addToast = useToastStore((state) => state.addToast);
+
     const handleDelete = async (assetId: number) => {
-        deleteAsset(assetId).then(() => mutate());
+        try {
+            await deleteAsset(assetId);
+            await mutate();
+            addToast({
+                variant: 'normal',
+                title: '삭제 완료',
+                description: '선택한 항목이 삭제되었습니다.'
+            });
+        } catch (error) {
+            addToast({
+                variant: 'critical',
+                title: '삭제 실패',
+                description: error instanceof Error ? error.message : '알 수 없는 오류가 발생했습니다.',
+            });
+        }
     };
 
     const handleEdit = (assetId: number) => {
         setSelectedAssetId(assetId);
         openModal('edit');
-    };   
+    };
 
     const AssetData = useAsset(data || [], handleDelete, handleEdit);
 
     const handleDeleteSelected = async () => {
-        if(selectedAssets.size === 0){
-            return alert('삭제할 항목을 선택해주세요.');
+        const isConfirmed = window.confirm('정말 삭제하시겠습니까?');
+
+        if (!isConfirmed) return;
+
+        if (selectedAssets.size === 0) {
+            addToast({
+                variant: 'warning',
+                title: '선택 필요',
+                description: '삭제할 항목을 선택해주세요.'
+            });
+            return;
         }
-        try{
+
+        try {
             await Promise.all(
                 Array.from(selectedAssets).map(asset => handleDelete(Number(asset.id)))
-            )
+            );
 
-            alert(`${selectedAssets.size} 개의 항목이 삭제 되었습니다.`);
+            addToast({
+                variant: 'normal',
+                title: '일괄 삭제 완료',
+                description: `${selectedAssets.size}개의 항목이 삭제되었습니다.`
+            });
             setSelectedAssets(new Set());
 
-        } catch (error){
+        } catch (error) {
+            addToast({
+                variant: 'critical',
+                title: '삭제 실패',
+                description: error instanceof Error ? error.message : '알 수 없는 오류가 발생했습니다.',
+            });
             console.error('삭제 중 오류가 발생했습니다.', error);
         }
     };
@@ -52,7 +89,7 @@ export default function AssetPage() {
                             className='bg-destructive-150 text-destructive-700 font-semibold hover:bg-destructive-200'
                             onClick={handleDeleteSelected}>삭제</Button>
                 </div>
-                {error && <StateInfoWrapper preset="defaultError" />}
+                {error && <StateInfoWrapper preset="defaultError"/>}
                 {isLoading && <Skeleton className="w-full h-100"/>}
                 {!isLoading && !error && AssetData && (
                     <DataTable
@@ -70,7 +107,7 @@ export default function AssetPage() {
                             );
                         }}
                     />
-                )} 
+                )}
             </div>
             <AssetRegistModal
                 isOpen={isOpen}
